ArticleWithUserRepositoryを作成
