+++ app/Controller/TopPageController.php
@@ -6,6 +6,7 @@ use App\Http\Request;
 use App\Http\Response;
 use App\Model\Category;
 use App\Repository\ArticleRepository;
+use App\Repository\ArticleWithUserRepository;
 use PDO;
 use Exception;
 
@@ -30,8 +31,8 @@ class TopPageController implements ControllerInterface {
      * @throws Exception
      */
     public function __invoke(Request $req, PDO $db): Response {
-        $articleRepository = new ArticleRepository($db);
-        $articlesWithUser = $articleRepository->getAllArticlesWithUser();
+        $articleWithUserRepository = new ArticleWithUserRepository($db);
+        $articlesWithUser = $articleWithUserRepository->getAllArticlesWithUser();
         $allCategories = Category::getAllCategories();
 
         ob_start();
